requirements.txt

bitarray==3.7.2
bitstring==4.3.1
black==25.9.0
cffi==2.0.0
click==8.3.0
cryptography==46.0.3
esptool==5.1.0
flake8==7.3.0
iniconfig==2.3.0
intelhex==2.3.0
markdown-it-py==4.0.0
mccabe==0.7.0
mdurl==0.1.2
mypy_extensions==1.1.0
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
pycodestyle==2.14.0
pycparser==2.23
pyflakes==3.4.0
pygame==2.6.1
Pygments==2.19.2
pyserial==3.5
pytest==8.4.2
pytokens==0.2.0
PyYAML==6.0.3
reedsolo==1.7.0
rich==14.2.0
rich-click==1.9.4
uvloop==0.19.0 ; sys_platform == 'linux'
websockets==12.0
